# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.0",
    "title": "Shop API",
    "description": "API testing for the shop"
  },
  "host": "localhost:3001",
  "tags": [
    {
      "name": "store",
      "description": "Inventory queries"
    },
    {
      "name": "user",
      "description": "User information"
    }
  ],
  "paths": {
    "/user/create": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create a user",
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User created"
          },
          "400": {
            "description": "Username already exists"
          }
        },
        "consumes": [
          "application/x-www-form-urlencoded"
        ]
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Login to user",
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Valid password"
          },
          "400": {
            "description": "Invalid password"
          }
        },
        "consumes": [
          "application/x-www-form-urlencoded"
        ]
      }
    },
    "/user/verify": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Verify JWT once user has logged in",
        "description": "",
        "security": [],
        "responses": {
          "200": {
            "description": "Valid"
          },
          "400": {
            "description": "Invalid"
          }
        }
      }
    },
    "/shop/create": {
      "post": {
        "tags": [
          "store"
        ],
        "summary": "Add item to shop",
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "category",
            "in": "formData",
            "required": true,
            "type": "string",
            "enum": [
              "PS4",
              "xBox",
              "PC"
            ]
          },
          {
            "name": "price",
            "in": "formData",
            "required": true,
            "type": "number",
            "description": "in pence e.g. 100 = £1.00"
          },
          {
            "name": "image1",
            "in": "formData",
            "required": true,
            "type": "file"
          },
          {
            "name": "tags",
            "in": "formData",
            "required": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "hidden",
            "in": "formData",
            "required": true,
            "type": "boolean"
          },
          {
            "name": "author",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Item added"
          },
          "400": {
            "description": "Item not added"
          }
        },
        "consumes": [
          "multipart/form-data"
        ]
      }
    }
  }
}